3.0: Modify imagebuilder extra chef attributes * To align with ExtraChefAttributes in cluster, this parameter accepts json format {"cluster": {"key": "value"}"} extra chef attributes * Supports chef attributes besides cfncluster, such as nfs, openssh etc.
